+++ examples/utils/migration-tools/src/migrator/migratorEntryPointPiece.ts
@@ -13,7 +13,6 @@ import { type ISimpleLoader } from "../simpleLoader/index.js";
 
 import type { DataTransformationCallback, IMigratableModel } from "./interfaces.js";
 import {
-	getModelFromContainer,
 	Migrator,
 	type LoadSourceContainerCallback,
 	type MigrationCallback,
@@ -37,6 +36,24 @@ async function getDataStoreEntryPoint(
 
 	return entryPointHandle.get();
 }
+
+/**
+ * Helper function for casting the container's entrypoint to the expected type.  Does a little extra
+ * type checking for added safety.
+ */
+const getModelFromContainer = async <ModelType>(container: IContainer): Promise<ModelType> => {
+	const entryPoint = (await container.getEntryPoint()) as {
+		model: ModelType;
+	};
+
+	// If the user tries to use this with an incompatible container runtime, we want to give them
+	// a comprehensible error message.  So distrust the type by default and do some basic type checking.
+	if (typeof entryPoint.model !== "object") {
+		throw new TypeError("Incompatible container runtime: doesn't provide model");
+	}
+
+	return entryPoint.model;
+};
 
 /**
  * Make a typical migration callback.
