+++ examples/utils/migration-tools/src/simpleLoader/index.ts
@@ -5,4 +5,4 @@
 
 export type { ISimpleLoader } from "./interfaces.js";
 export { SessionStorageSimpleLoader } from "./sessionStorageSimpleLoader.js";
-export { SimpleLoader, waitForAtLeastSequenceNumber } from "./simpleLoader.js";
+export { SimpleLoader } from "./simpleLoader.js";
